projects/rag/server/pyproject.toml
[tool.poetry]
name = "rag"
version = "0.1.0"
description = "rag repository requirements"
authors = [
    "jjonhwa",
    "jun048098",
    "kyhoon",
    "Seonu-Lim",
]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
faiss-cpu = "^1.8.0.post1"
fastapi = "^0.112.1"
gunicorn = "^22.0.0"
kiwipiepy = "^0.18.0"
langchain-community = "^0.2.7"
langchain-huggingface = "^0.0.3"
langchain-openai = "^0.1.17"
langgraph = "^0.2.12"
langserve = "^0.2.2"
rank-bm25 = "^0.2.2"
rich = "^13.7.1"
typer = "^0.12.3"
typing-extensions = "^4.12.2"

[tool.poetry.group.dev.dependencies]
black = "24.4.2"
isort = "5.13.2"
pre-commit = "^3.7.1"
pytest = "^8.3.2"
pytest-asyncio = "^0.24.0"
ruff = "0.5.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "DEBUG"
